Make inithost.sh repeatable.

+++ inithost.sh
@@ -11,20 +11,44 @@ apt-get update && apt-get install -y \
   apt-utils \
   git
 
-groupadd -g 1700 ${OS_USER}
-read -p "Enter password for OS user ${OS_USER}: " GISADMIN_PASSWORD
-useradd -u 17000 -g 1700 -d ${USER_DIR} -m -s /bin/bash -p $(echo ${GISADMIN_PASSWORD} | openssl passwd -1 -stdin) ${OS_USER}
+if [ $(getent group ${OS_USER}) ] ; then
+  echo 'Group ${OS_USER} already exists.'
+else
+  echo 'Create group ${OS_USER} with id 1700.'
+  groupadd -g 1700 ${OS_USER}
+fi
 
+if [ $(getent passwd ${OS_USER}) ] ; then
+  echo 'User ${OS_USER} already exists.'
+else
+  echo 'Create user ${OS_USER} with id:17000 and add to group ${OS_USER} gid:1700.'
+  read -p "Enter password for OS user ${OS_USER}: " GISADMIN_PASSWORD
+  useradd -u 17000 -g 1700 -d ${USER_DIR} -m -s /bin/bash -p $(echo ${GISADMIN_PASSWORD} | openssl passwd -1 -stdin) ${OS_USER}
+fi
 cd ${USER_DIR}
 
+if [ -d ./kvwmap-server ] ; then
+  echo 'Stop kvwmap-server and uninstall all.'
+  dcm uninstall all
+fi
+
+if [ -d ./kvwmap-server ] ; then
+  echo "Abort by user."
+  exit
+fi
+
+echo 'Clone kvwmap-server repository to ./kvwmap-server.'
 git clone https://github.com/pkorduan/kvwmap-server.git
 
 kvwmap-server/dcm install kvwmap
 
 source ~/.bash_rc
 
-sed -i -e "s|#PermitRootLogin prohibit-password|PermitRootLogin no|g" /etc/ssh/sshd_config
-
+read -p "Enter port for ssh login: " SSH_PORT
+sed -i \
+    -e "s|#PermitRootLogin prohibit-password|PermitRootLogin no|g" \
+    -e "s|#Port 22|Port ${SSH_PORT}|g" \
+    /etc/ssh/sshd_config
 /etc/init.d/ssh reload
 
 dcm run pgsql
